// geo-brand-visibility-be-dev/src/prompt/dto/get-responses.dto.ts
import { ApiProperty } from '@nestjs/swagger';

export class ResponseModelDTO {
  @ApiProperty({
    description: 'AI Model ID',
    type: 'string',
    example: 'f6a1b2c3-d4e5-6789-abcd-ef0123456789',
  })
  id: string;

  @ApiProperty({
    description: 'AI Model Name',
    type: 'string',
    example: 'gpt-4o',
  })
  name: string;
}

export class ResponseCitationDTO {
  @ApiProperty({
    description: 'The URL of the citation',
    type: 'string',
    example: 'https://openai.com/research',
  })
  url: string;

  @ApiProperty({
    description: 'The title of the citation',
    type: 'string',
    example: 'OpenAI Research',
    nullable: true,
  })
  title: string | null;

  @ApiProperty({
    description: 'The domain of the citation',
    type: 'string',
    example: 'openai.com',
  })
  domain: string;
}

export class GetResponsesResponseDTO {
  @ApiProperty({ example: 'b123f7aa-12ef-45cd-8b19-123456789abc' })
  id: string;

  @ApiProperty({
    description: 'The generated response text',
    type: 'string',
    example: 'This is the answer generated by the model.',
  })
  response: string;

  @ApiProperty({
    type: [String],
    description: 'List of questions that this response is related to.',
    example: ['What are AI tools?', 'Top AI research trends'],
  })
  relatedQuestions: string[];

  @ApiProperty({
    type: ResponseModelDTO,
    description: 'Information about the AI model used to generate the response',
  })
  model: ResponseModelDTO;

  @ApiProperty({
    type: [ResponseCitationDTO],
    description: 'Citations used in generating the response',
  })
  citations: ResponseCitationDTO[];
}